More CMake changes
All algorithms that can be built with CMake are now found automatically. There is no need to maintain a list in the CMakeLists.txt file in the algorithms/ folder.

On Unix systems, the search process is highlighted in blue. Oh, so beautiful!

I also made my first CMake macro! It mostly came from here though:
https://stackoverflow.com/questions/18968979/how-to-get-colorized-output-with-cmake